+++ lib/payment_server_web/types/balance.ex
@@ -2,7 +2,7 @@ defmodule PaymentServerWeb.Types.Balance do
   use Absinthe.Schema.Notation 
 
   object :balance do 
-    field :amount, :float
+    field :amount, :integer
     field :currency, :currency
     field :timestamp, :string
   end
